+++ NEWS.md
@@ -1,4 +1,8 @@
-# lsatTS changelog
+# LandsatTS changelog
+
+## 1.0.2 (released: 30 November 2022)
+- Changed package name from lsatTS to LandsatTS.
+- Added function lsat_calibrate_poly().
 
 ## 1.0.1 (released: 2 June 2022)
 - Bug fix to address error in lsat_get_pixel_centers() caused by change in the URL for the WRS2 geometries hosted by USGS. See Issue #35: https://github.com/logan-berner/lsatTS/issues/35